dynamic keybinding work, meta-button safeguard
initially, if a number of button presses occur without any of them being
a meta-key - we assume that the layout does not match the current
keyboard. This happens with internationalization issues, input platform
switching etc. In that case, a bar pops up that allows you to (at the
moment, temporarily) rebind the keys used for meta_1 / meta_2. This code
will soon be used for all input binding / rebinding.