inreased debugging verbosity, lbar bug work
interesting engine bug that was assumed to be connected with durden:
when returning from a VT switch, a lot of visual state changes seemed to
be stale, and lbar etc. didn't update. It seems like >something< in the
graphics state machine gets messed up on VT restore that takes some
attach / detach event until things return back to normal (!?). Try with
imagetest, load a few images, switch VT back and forth then load again.
Changes are not updated. This one can be really hairy, especially if the
problem comes in the egl-dri layer.