switch rendertargets in display iterators
This is an easy source of bugs for multidisplay that is worth
keeping track of, as the display management relies on changing the
global state in the default rendertarget attachment, any operation
that will force window/tiler etc. need to have the specific display
activated or the items will start appearing on the wrong screen.