SDL/simple mode workaround
This bug is somewhat interesting. We had a fillrate- problem
with the main rendertarget having secondary displays drawn to it
which was fixed with by never showing them, just mapping.
This lead to the case where the SDL platform (which fails on
mapping alternate rendertargets to main output etc. as it is
not implemented in the platform in arcan). The workaround while
waiting for the platform- to be fixed in arcan is to probe initial
mapping and if it fails, show the main rt anyway.