defer window-ws attachment
Previously, as soon as the window vid was available, it was added
to the workspace. This had the drawback of flickering / bad resizes
when the window was immediately terminated etc. Now we wait until
the first resize call (which always happens with the first delivered
frame). This has the effect that there is a fully qualified window
still living even though it's not attached to a workspace, so it is
likely that there are some "wnd.parent, wnd.space" assumptions left.

FossilOrigin-Name: c544dcf61361c441946bf9a153eff2acfd16f0f8aec61669120ba7df3a260853