(autolayouter) fix selection-bug, refactoring..
the annoying 'why did my selected window disappear' seem to have
been pinpointed to an asynch race when deleting windows in the side
columns, so on the re-eval/resize we simply make sure that there
is an active selection.

Slight rework of the registration/setup so we can add a new layout
mode, 'book'.