fullscreen mouse coordinate- management fix
Fix relies on a bump in arcan build to move to api version 11.

This one managed to hit a big one across the board which may explain quite some
problems and inconsistencies with mouse support. Spec-wise, the mouse input
table should pack % 2 = 0 sample values as relative or absolute based on if the
relative boolean was set or not. Both mouse.lua and the sdl/evdev platforms
failed to follow this "sortof". The reason it was masked for so long seem to be
the repacking done in most WM scripts used.

FossilOrigin-Name: 8ff69b119e245f007df5a3caca3c43645fc6b9cd141a1415995b3f8c61791188