new tool, (overlay)
This goes well with window/slice when you want to keep track of a
small part of a window without interrupting your normal workflow.
It places a non-interactive chain of images that grows down from
the top-right or top-left corner where the image is a copy of the
contents of the targeted window.

FossilOrigin-Name: f9a02e0283161760ae7779d575f12ce3a7006ccf5db6f03592a74592505d4bff