new tool, (overlay)
This goes well with window/slice when you want to keep track of a
small part of a window without interrupting your normal workflow.
It places a non-interactive chain of images that grows down from
the top-right or top-left corner where the image is a copy of the
contents of the targeted window.