add initial support for terminal groups
This is a complex feature with many subtle cases and not likely to work 100% at
first go. Each window can now carry a set of alternates, i.e. hidden windows
that can be swapped into the hierarchy in place of the current one. A terminal
group is a set of clients connected via a terminal-unique connection point, and
each new connection via this path gets force-added as an alternate to the
terminal. This allows a graphical client to be launched from the terminal and
hide the terminal in the background until deleted or swapped in.