(tools/advfloat) let background receive input
This is a special little thing. Via target/window you can set the
window external source as the background wallpaper. If running in
floating layout mode and all windows are deselected, any input on
the background will be forwarded to the external target. This let
you run an Xorg or VM instance as you 'wallpaper' and input will
simply fall-through if nothing else will receive it in beforehand.