mouse - emit release events on press-sources
The previous approach to this was that you only got the button event
that occured on your surface id, meaning that some clients would never
get a 'release' event to match a 'press' event. This is annoying for
some clients, particularly xarcan. This affects popups in wayland
clients as well, so that will need a separate fix.