wayland-toplevel, popup position and maximize fix
list of issues is steadily shrinking, the autocrop on border should be
working now, and the maximize should actually account for the statusbar
being at the top of the screen.

there are still a number of issues with the tiling mode, the sizing
of reparented toplevels, impostor input forwarding and possibly some
regressions in animations and autolayout for some scale modes.