Big refactor of the menu system++
This is an attempt at harmonizing the different namespaces and browsing
and a big cleanup to the mess that is the interplay between bindings,
the menu and the various fglobal/dispatch/etc. functions. The big thing
is that the iopipes are now an ipc system that uses a socket, the
special ! and # namespaces are dropped in favor of a global tree that
also includes t he browser itself.

FossilOrigin-Name: 89ed3b1ccdfc771e954976bc43bb9b606cc2b2025cb120637dfed577158fd6df