refactor/cleanup work for notification system
The code isn't 100% complete on this one yet, but it is quite close.
What this adds is some cleanup to how widgets are created and freed,
along with a notification widget that subscribes to client and system
notification events.

The first test-case for this is to make sure that the relevant message
from the last crash on crash recovery gets added as a critical
notification so it gets easier to act upon it.