add fatal event handler
In synch with the arcan-master changes of adding a handler for script
errors to permit the script one last chance to save valuable user data,
we set it to expose a stack trace and the last exposed dispatch symbol.

This should catch most of the possible failure sources in quick-to-debug
ways.