rework ipc/monitor mode, consolidate logging
This is quite a big but important change in the setup of things. Most
of the various subsystems are now hooked up to the monitor mode, and
the code for adding new ones has been streamlined.

The changes to display.lua also bring in the first round of refactoring
trying to right the wrongs that come with display profiles and the
primary display in its various stated. This ties into one of the most
complicated subsystems (switch VTs, hotplug all/none/some displays,
reset, ...) that definitely won't be 100% this version push but it
is one of the few last bandaids that need to be ripped off before we
can move to greater heights.

FossilOrigin-Name: 01cfcc6b893c1ac06b8606a0c432ad6e9ada077a030a9b76ee4d819bacfaf7b6