(shaders) added rounded/shadowed effect shader
This just repeats the setup of the rounded shader that is used in some
of the UI components now. The added changes to shadermgmt is to allow
the global/config/visual/shaders menu setup of the fx- shaders, but not
pay the cost of the rendertarget redirect that they use.