add controls for rendertarget-clearcolor
This allows the default clear- color for display rendertargets to be
set. This is helpful when a wallpaper is not wanted, but some kind of
tint to transparent windows, decorations and padding is still warranted.

FossilOrigin-Name: a0ef89a1dd2f1589f01bc39fbbc5e9534036d6fffea82ad3067aa5072aa8494e