remove status/output fifo
This drops the option to parse and deal with external fifo- text
controls for adding contents to the statusbar. It will be swapped for a
tool that exposes a connection point for docking into the external area
of the bar instead.

FossilOrigin-Name: 7926a21a8bf3703dc9e2445a65594aeab4625580854630b54da0c167a837e8c3