(workspace) add manual attachment control
This lets you set the default workspace where new windows will spawn.
Two pieces are missing from this feature still. One is a visual
indicator in the statusbar that a window was added to the workspace in
question. The other is to specify additional rules that would pre-set
the designated workspace for a window. Most of this infrastructure is
already in place as part of crash recovery, but having custom rules
helps type-arranging.

FossilOrigin-Name: aacac4a46f64a0cb82aa6268618a9eff342c5279d1143e75820c923b61b778bb