add draft 'streamdeck' tool
This tool allows you to interface with 'input-display' hybrid devices
like the 'elgato streamdeck'. There are some minor kinks left as for
rebuild criteria and generation of text-icons, but the overall concept
works.

It is exposed as /global/tools/streamdeck and the pattern is to open
a connection point (streamdeck/define=name), then specify the resources
that should be mapped to regions on the display, and their respective
priorities. When done, run streamdeck/name/open to activate.

The different groups currently:
- windows : current workspace windows, icons (if defined) or canvas
- labels : exposed input bindings for the currently selected window
- titlebar : icons bound to the current selected window decorations

Missing but incoming:
- workspaces : preview for each present workspace on the active display
- custom : bind custom input/path

An example of a driver that can be used with this tool can be found at
https://github.com/letoram/arcan-devices