tools - add todo tool
This one has been laying on the desk for a long while, as it is normally
how I arrange what I am working on depending on which context I am in.

It is a quick todo tool which exposes itself as a statusbar button. The
general usage is:

/global/tools/todo/set_group=work
/global/tools/todo/enable/statusbar=left
/global/tools/todo/add/item=demo:prepare a demo for xyz

group can only be modified when the tool isn't enabled, so switching:

/global/tools/todo/disable
/global/tools/todo/set_group=pleasure
/global/tools/todo/add/item=wax:taint needs a good ol' trimming
/global/tools/todo/enable/statusbar=left

the button label shows the shortname, and clicking will yield a popup
where the item can be marked as completed or postpone it.

the items in the current group can also be accessed via:

/global/tools/todo/items/current (current active item)
/global/tools/todo/items/last (last added item, if known)
/global/tools/todo/items/wax/complete (or by name)

A quick and dirty thing but quite useful, with the control/IPC interface,
translating git- tickets etc. and a focus thing in just glancing at the
titlebar to see what I was supposed to be doing.