(tools/vrviewer) drop the vrviewer tool
The approach of sharing code with safespaces in this way was quite
cumbersome to work with and slowed down the development of both sides.
The feature will be re-introduced after some changes to arcan where _lwa
clients that enable the vr- substructure to receive and map hmd-l/ hmd-r
with direct scanout.