always send display-table on display-hint
This should close down a number of cases where subwindows failed to
use the proper density of the display it is attached to. Basically
this is due to the _acquire vs _connect semantics in shmif not being
able to provide an initial table of display properties (though that
should probably be fixed in shmif itself to retain all non-font md
as a shared singleton).

FossilOrigin-Name: c08e3e1fdc87e8514bdcca240a5388b98a610f3a6b4475d049f626fe577c24d2