mouse-warp on window creation
This fixes an issue where the mouse would warp to 0,0 on the spawning
of a new selected window on the workspace. The solution is not very
nice, i.e. ignoring the warping control if the window is selected in
the same clock-tick as when it is attached, but tracking the selection
source in other ways was even messier.