popup - add support for preset field in menus
this is a bit of an experiment, but we need more fields to allow
non-keyboard input to valuetypes. presets is like set for such
cases, other option is to provide a range and stepping function,
but get to that later.

FossilOrigin-Name: 22d08bf45d62b3fb344a695954fbcc44f6dae9e37d4bf8ea4b39b07ceb0ac640