iostatem/touchm refactor stage 1
This should greatly improve the behavior of some touch displays,
especially for the absmouse classifier. This refactor also makes
the handover/registration between iostate and touch more explicit
to allow other input devices and tools to reuse more infrastructure.

Next stage to this is splitting touchm into the generic management and
the classifiers, so that it becomes easier to define a new classifier,
as well as export a menu entry to dynamically swap between classifiers.