add draft profile-picker tool
Still incomplete, but need to synch between machines. This tool will
replace the meta-key binding guard on first-run. Eventually it should
evolve into a 'welcome/first-config' tool, and intermediate steps
provide a chain of popups to set default schemes for:

 * window-management
 * keybindings
 * colors
 * security level

The follow-up is adding schemes for the common defaults of popular
WMs so it is little to no effort 'getting into' using durden for the
first time.