rename xwl- surfaces
Some confusion with the naming seems involved in the bits that the
xwayland surfaces do not get properly restored on reset or crash
recovery. This won't address that, but it is part of it.

FossilOrigin-Name: 49ef0805b657342493a8e13968345d1de0155dc8580e6f1f0f109bd0e9d4bde7