external connections getting disabled
The rate-limit implementation and incoming-endpoint tracking needs
a real refactor / makeover as a helper that can be used for all the
different places we do connection point allocation from.

During some edge-cases with instant-terminated connections we get
events that fail to trigger external connection point rebuild, this
patch should help with that.

There is still some -to be determined- condition that will trigger this
issue, as well as a possible rate-limit stall on wayland clients in
particular due to how their allocations work.