implement more of layout-based size suggestion
During the pre-roll stage the WM has an option of providing a hint as to
what the starting dimensions should be. These have been fixed at a bad
resolution (300x300) for debugging reasons, and it is time to switch it
to more proper methods.

The support for the different modes are still somewhat validated, and
the current design can't account for tiling specifically as we need to
test-insert a window at the specific point, run a resize round that does
not commit and use the value we get from there.