cli_group - allow target/window/swap/swallow/first
This adds a toggle between the cli_group window swallow slot and its
parent. It also fixes an issue with the hidden windows and tree- slot
not being matched, and swallowed windows not migrating/reassigning.

Should the feature play out well, expand the implementation to allow
multiple windows in the swallow slot and an action to swallow other
attached windows.