wire in partial wayland data-device support
this takes care of registering type providers (recall data-dev in
wayland exposes a set of mime types that the recipient can pick from),
as well as a test/fast-path for utf8- treating it the same as other
short-text from a regular arcan client.

more menu mapping is needed for the basic feature, and some changes
to upstream arcan-wayland, but most of the interface should be figured
out by now.