override 'native' default font density
This is an old relic from how we did mixed DPI before the proper engine
changes had been made. With changes in upstream arcan, the proper values
for primary display HPPCM will be set immediately >and for the internal
font management<.

This caused weird padding issues with sbar and tbar specifically, as
well as comically oversized icons on HDPI. The real fix is that the font
sizes should never have any transforms, proper density values on rtgts
and correct button rebuild-state in the button bar widget, until then,
fake revert to 38.4 ppcm.