icon sizing propagation fix, titlebar pos-control
This continues the 'statusbar at sides' with doing the same for the
titlebar. There still is some issue with stacking mode, the bar on the
right and maximizing and probably others. This should also smoke out
some of the multi-display migration issues causing some icons to have
the wrong size or attachment, but it is a complex issue that needs
more testing/work.

FossilOrigin-Name: 4a70894ad6fc8576f884e194784dffec9adcb67a70feb3b6b1f3d1f0646975c7