ensure correct rendertarget on 2nd display
This should fix the elusive problem of some statusbar icons being
'wrong' when new displays are plugged in. The cause is a recurring
design problem in durden due to its history of driving multidisp api
design for the engine itself.

Since we are not explicit in which rendertarget we attach the new
vids to, a global 'current_attachment' is used, but this might not
be the intended display. Most of the time window creation is tied
to the currently active display and is then forcibly migrated
elsewhere and this problem does not exist. For certain conditions,
e.g. density changes - the statusbar and its icons are rerastered
to match the new target density. This will cause the creation of
icons intended for a display that isn't the current attachment.

The solution is to ensure that the intended rendertarget is the
current active one, but it is not easy on long event-chains.

FossilOrigin-Name: 24cf568526a112d09af04273a483a152e173b146f7a47481f311bc95c29f4469