uiprim/bar - nested state on reset workaround
There is an 'interesting' bug when the titlebar is hidden and set to
merge into the statusbar. After reset, the 'nested' state for adopted
windows seem to be somewhat broken (the text appears but the selection
state does not). Destroying the window in this case had a null-deref
in parent:set_... so the parent state seem to be broken on destroy.

Check for this case for now so the scripting error goes away, but it
needs a proper fix.