remove 'simple' mode and allow per-display direct
Previously the 'simple' mode disabled a lot of composition options in
order to save memory bandwidth on limited platforms like many ARM SBCs.
With the engine refactoring that has been done to support more diverse
colour composition paths (HDR) there is a better option to have the same
optimization.

This will need to be reflected in some documentation changes as well,
but to get the direct performance increase now, it has to be set as the
default in engine via arcan_db or enabled through map_video_display
with the HINT_DIRECT flag.