add compact/sidepad to tbar/sbar
This is a step towards allowing a minimal statusbar that more fits
the behaviour of the doc/tray in windowManager and friends. The
same goes for titlebars that have the compact BeOS look.

There are some quirks to this that isn't entirely worked out but a
fair point to commit/pause anyhow.

FossilOrigin-Name: 6f6840479f7cf4f0b1dba46ea8586f1ec6fd65232b6060f017b944fb7bb5d1be