add compact/sidepad to tbar/sbar
This is a step towards allowing a minimal statusbar that more fits
the behaviour of the doc/tray in windowManager and friends. The
same goes for titlebars that have the compact BeOS look.

There are some quirks to this that isn't entirely worked out but a
fair point to commit/pause anyhow.