respect viewport- embedding requests, fix spawnctl
external clients can request a viewport embedding wherein a window turns
into a subsegment. This is used by both some tui and some wayland
clients. Previously this was ignored, now the surface is added as a
subsurface and with the option of 'mouse-drag' detaching.

spawnctl 'draw to spawn' mode did not clamp max dimensions correctly,
causing a subsequent resize to ignore the drawn dimensions.