add basic picking from user-namespaces
with arcan 0.6.1.1+ user namespaces can be defined for read only or
read-write of paths that appear and disappear at run-time, such as
something managed by a mounting service.

this patch tests if this feature is available, and extends the menu
and preview generation to follow the semantics of the namespace api.

There are some issues with this still, the easiest to work from is
that of picking a workspace background which will not get to the right
path and cancel-out when trying to traverse upwards.

The other parts that need more work is sending a file to a window,
or clipboard, as well as copying from a file or window, filtering and
showing namespaces based on access permissions.