fix cursor jumping on wl-x11 window popup
This fixes a bug where the mouse cursor seemed to jump around
when a wl-x11 popup window was created. It is about the 'mouse
follows last known input focus' feature that is mostly great for
keyboard nav. and you want to have the cursor where it was last
for any one window.

The problem was with the proxy x11 window that had to be created
in advance before knowing type / purpose. When that window became
a popup (which is assigned a parent and not treated as a normal
window), the proxy was destroyed, triggering the auto-warp from
its previous parent.

The quick workaround is to just save the mouse coordinates before
destroying the proxy, and then warp back.

FossilOrigin-Name: 1876a9fc792105a8fb4ef0ba1bacaf879fa9299095de7c44970f53b3a0145c6e