sbar-hud visibility
This fixes 2/3 of the current issues with the statusbar acting
strangely when attached to the hud. The buttons still does not
reattach cleanly, but at lest the background and sizing is back
to normal.

Also adds a highlight with meta-over + meta-drag for embedded
surfaces slated for decoupling.

FossilOrigin-Name: 32b1c975eed8ac2cd4329279123198039c1d71ec6e9870c1435c4ecbec65da56