atype/terminal: route 'message' as notification
This was an old remnant from a hack on afsrv_terminal where the cell
dimensions were being forwarded as a message - something that the
notification facility in TUI relies on and we now have tui clients
that actually use it.

This behaviour was practically deprecated when server-side text was
added as that information now comes from the resizing/viewport
events.

FossilOrigin-Name: 78ce351b9561034fc30e5e808615c39a4a2469f230d81251085e31604d0a2700