shadow regression fix, log-fmt error
The refactor from border shadow into suppl code attempted to use
the scale mode property of image linking in a way that either hints
at a bug in the engine or how we actually size the window. It also
failed to set the unpickable mask causing some windows to not only
look gross but also be hard to move. Reverted to the old-ish setup.