work torwards better x11 client integration
This continues the previous big batch on changes needed to atypes/x11
in order to either let Xarcan + wm completely control a single workspace
including arcan clients that spawn/reassign to the workspace or to let
us steal away select or all toplevel windows. It is still incomplete
and work in progress.