tool/a12 - start adding button and discovery ctrls
This is a bit weird to develop as it hinges on changes throughout
the stack as well as experimental durden branches with unmerged
code.

This adds the basics for controlling buttons added to the statusbar
when signalling changes to a discovered directory server or known
sources. (sinks come with virtual displays and should integrate in
the other sharing menus, including 'sinking' to a directory').

The real kicker for the coming release is to get dynamic updates for an
established directory server when there is a new source or appl
available, or reacting when an appl that is running has changed, with
the default being state-store + reload + state-restore.

The later would allow a live-coding scenario where someone with access
could push an appl or changes to a shared directory, and others using it
immediately update, with crash/bug collection and
revert-to-last-working.

FossilOrigin-Name: d26bc42cf1c737c55fd589b908438bb1aa5ed42566b5d99e75134bd8b3f1550b