atypes/x11 - more clipboard fixes
Excluding one painpoint (actual input trigger for 'paste' activation)
this should have working selection buffer update and paste.

It is unclear if the synch should be made on the xarcan side or on the
wm side as there is no sure-fire way to guarantee that a certain input
yields a paste action in a client (though convention on ctrl+v /
midbtn). The safe approach to this requires quite a lot of tracking to
provide a fair migration path, i.e. if something yields /target/paste
while other input is still pending.

An option would be to add a virtual keyboard / sink in the Xarcan side
that runs a known keymap and only receives input from the pasteboard,
which is a complicated way to go about things. Maybe there is something
in the accessibility / IME troves?

FossilOrigin-Name: 7f7357328ecf8fc325d05fa5243ae0d83e608289d8c4f01c64f3761e91d97f12